♻️ refactor(rules): split validation.rs into domain-specific checkers (#92)
## Why (User Story)
As a developer, I want rule validation organized by validation domain so
that schema, schedule, notification, and filter checks are independently
testable.

## What Changed
- Split 1158-line validation.rs into 6-file module:
  - mod.rs: ValidationResult, ValidationError/Warning types, validate_rule()
  - schema_checks.rs: Schema and detection validation
  - schedule_checks.rs: Cron expression validation
  - notification_checks.rs: Notification config validation
  - filter_checks.rs: Filter expression validation
  - fuzzy.rs: Fuzzy string matching for "Did you mean?" suggestions

Related to #96
